More type consolidation

+++ lib/MCP23X17/MCP23X17.cpp
@@ -61,11 +61,11 @@ void MCP23017::begin(uint8_t I2C_address, uint8_t interrupt_pin)
 /**
  * @brief Reads the state of the GPIO pins
  * 
- * @return unsigned int state of the GPIO pins on port A and B in one. B on the MSB
+ * @return uint16_t state of the GPIO pins on port A and B in one. B on the MSB
  */
-unsigned int MCP23017::ReadGPIOs() 
+uint16_t MCP23017::ReadGPIOs() 
 { 
-    unsigned int gpio = 0;
+    uint16_t gpio = 0;
     Wire.beginTransmission(_device_address);
     Wire.write(MCP23017_GPIOB);
     Wire.endTransmission();
@@ -83,16 +83,16 @@ unsigned int MCP23017::ReadGPIOs()
 /**
  * @brief Gets the interrupt flag registers value, on which set bits means that a pin caused an interrupt
  * 
- * @return unsigned int containing the interrupt flag registers value with the port B on the MSB
+ * @return uint16_t containing the interrupt flag registers value with the port B on the MSB
  */
-unsigned int MCP23017::ReadIntFlag() 
+uint16_t MCP23017::ReadIntFlag() 
 {
     Wire.beginTransmission(_device_address);
     Wire.write(MCP23017_INTFB);
     Wire.endTransmission();
     Wire.requestFrom(_device_address, 2);
     uint8_t i = 0;
-    unsigned int intflag = 0;
+    uint16_t intflag = 0;
     while (Wire.available())
     {
         intflag = intflag << (8*i);
